pyproject.toml
[project]
name = "monday-grabber"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jean Boutros" }
]
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
    "jq>=1.8.0",
    "polars>=1.37.1",
]

[project.scripts]
monday-grabber = "monday_grabber:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/monday_grabber"]
exclude = [
    "*.json",
    "__pycache__",
]
